requirements.txt
google-cloud-aiplatform==1.127.0
google-cloud-storage
cloudpickle==3.1.2
pydantic==2.12.4